Basic functionality for lastPlayed via edit & add forms.

+++ js/common.js
@@ -46,44 +46,51 @@ function GetMissionAuthor(preSelect) {
     });
 }
 
-function CheckRights(obj, userid, ACL) {
+function CheckRights(rowid, userid, ACL) {
     var currentUser = Parse.User.current();
-    var rowid = obj.id;
 
-    // If current user has created the entry, add edit & delete options
+    // If the current user has created the entry, add edit & delete buttons
     if (currentUser.id == userid) {
-          $("#" + rowid).append(
-            '<ul><li><a href="edit.html?row=' +
-            rowid + '">Edit</a></li><li><a href="#" onClick="DeletePopup(\'' +
-            rowid +
-            '\')">Delete</a></li></ul>');
+      $("#" + rowid).append(
+        '<ul><li><a href="edit.html?row=' +
+        rowid + '">Edit</a></li><li><a href="#" onClick="DeletePopup(\'' +
+        rowid +
+        '\')">Delete</a></li></ul>');
+      //return;
     }
 
+    // Check if admin
     var query = new Parse.Query(Parse.Role);
     query.equalTo("users", currentUser);
     var value = null;
     query.find({
         success: function(roles) {
+            for (var x = 0; x < roles.length; x++) {
+                if (ACL.getWriteAccess(roles[x])) {
 
-                for (var x = 0; x < roles.length; x++) {
+                    // Add incrementor for play-counter
+                    //$('#' + rowid + '_cellPlayed').append(' <a href="#">+</a>');
                     
-                    if (ACL.getWriteAccess(roles[x])) {
+                    // If admin is not creator of the entry, add edit & delete buttons
+                    if (currentUser.id != userid) {
                         $("#" + rowid).append(
-                          '<ul style="float:right"><li><a href="edit.html?row=' +
+                          '<ul><li><a href="edit.html?row=' +
                             rowid + '">Edit</a></li><li><a href="#" onClick="DeletePopup(\'' +
                             rowid +
                             '\')">Delete</a></li></ul>');
-
-                        $('#'+rowid+'.cellPlayed').append(' +');
-                        return;
                     }
+                    
+                    return;
                 }
+            }
         },
         error: function(error) {
             console.log("Error: " + error.code + " " + error.message);
         }
+
     });
 }
+
 
 function MissionSaveError(string) {
     $("#errorEdit").text(string);
